PATCH: Fixed parameters and refactored method

+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/services/PostService.java
@@ -8,10 +8,10 @@ import java.util.Optional;
 
 public interface PostService {
 
-    Optional<Post> getPostById(int id);
+    Optional<Post> getPostById(long id);
     Post create(long userId, long debateId, String content);
     Post createWithEmail(String userEmail, long debateId, String content);
     List<Post> getPostsByDebate(long debateId, int page);
-    Optional<PublicPost> getPublicPostById(int id);
+    Optional<PublicPost> getPublicPostById(long id);
     List<PublicPost> getPublicPostsByDebate(long debateId, int page);
 }
